refactor: move logout button to settings header
