+++ frontend/src/components/layout/pageLayout/regions.tsx
@@ -2,10 +2,6 @@ import SelfProfile from "@/components/composite/selfProfile";
 import styles from "./pageLayout.module.css";
 import { User } from "@/types/user";
 import { ReactNode } from "react";
-import { Button } from "@/components/atomic";
-import { Api } from "@/util/api";
-import config from "@/config";
-import { useRouter } from "next/navigation";
 
 export interface RegionProps {
   className?: string;
@@ -21,19 +17,6 @@ export interface StickyProps {
 }
 
 export function Sticky({ start, children, end, self, className }: StickyProps) {
-  "use client";
-  const router = useRouter();
-
-  const handleLogout = async () => {
-    try {
-      await Api.post(`${config.api.baseUrl}/users/logout`);
-      router.push("/");
-    } catch (error) {
-      console.log(error);
-      // TODO: Display toast
-    }
-  };
-
   return (
     <div className={`${styles.sticky} ${className ?? ""}`}>
       <div className={styles.stickyContent}>
@@ -41,15 +24,7 @@ export function Sticky({ start, children, end, self, className }: StickyProps) {
         {children ?? <div />}
 
         <div className={styles.row}>
-          {end}{" "}
-          {self && (
-            <>
-              <SelfProfile key="profile" user={self} />
-              <Button variant="subtle" onClick={handleLogout}>
-                Log out
-              </Button>
-            </>
-          )}
+          {end} {self && <SelfProfile key="profile" user={self} />}
         </div>
       </div>
     </div>
