+++ frontend/src/components/settings/sectionTitle/sectionTitle.module.css
@@ -1,4 +1,5 @@
 .sectionTitle {
+  font-weight: 500;
   margin-block-end: var(--step-0);
 
   &:not(:first-child) {
