+++ frontend/src/components/settings/settings.module.css
@@ -1,5 +1,5 @@
 .header {
-  align-items: center;
+  align-items: flex-end;
   display: flex;
   flex-direction: row;
   gap: var(--step-2);
